Set int type to Uint64
This is the type used for accuracy_radius

+++ mmdb_enrich.go
@@ -18,7 +18,7 @@ func to_mmdb(object interface{}) mmdbtype.DataType {
 	case string:
 		return mmdbtype.String(t) // t has type string
 	case int:
-		return mmdbtype.Int32(t)
+		return mmdbtype.Uint64(t)
 	case float32:
 		return mmdbtype.Float32(t)
 	case float64:
